requirements.txt
PyYAML
elsa>=0.1.2
jinja2
markdown
Flask
